WIP, all tests passing on laptop

+++ DataField.hpp
@@ -17,18 +17,18 @@ public:
     // Reads the data field from the "buffer" memory location without
     // considering byte ordering.
     virtual unsigned long readRaw(std::uint8_t* buffer,
-                                  unsigned int  bit_offset = 0);
+                                  unsigned long bit_offset = 0);
 
     // Reads the data field from the "buffer" memory location, swapping if the
     // source byte order does not match the byte ordering of this field
     virtual unsigned long readRaw(std::uint8_t*   buffer,
                                   misc::ByteOrder source_byte_order,
-                                  unsigned int    bit_offset = 0) = 0;
+                                  unsigned long   bit_offset = 0) = 0;
 
     // Writes the data field to the "buffer" memory location without considering
     // byte ordering.
     virtual unsigned long writeRaw(std::uint8_t* buffer,
-                                   unsigned int  bit_offset = 0) const;
+                                   unsigned long bit_offset = 0) const;
 
     // Writes the data field to the "buffer" memory location, swapping at the
     // destination if the destination byte order does not match the byte
@@ -36,7 +36,7 @@ public:
     virtual unsigned long writeRaw(
         std::uint8_t*   buffer,
         misc::ByteOrder destination_byte_order,
-        unsigned int    bit_offset = 0) const = 0;
+        unsigned long    bit_offset = 0) const = 0;
 
     // Returns the size of this data field in bits.  This will equal the number
     // of bits written by writeRaw() and read by readRaw().
